idn: rename the files to idn.[ch] and hold all IDN functions
